Migrated to latest dependencies
Some math appears to have changed underneath so the tests needed to be
updated. I've added an inefficient test of the max hamming distance for
the comaparisons. I'm not happy with the dhash that is getting a hamming
distance of 4 on samples 03. However with 64 bits of information in the
hash, a hamming distance of 4 says that the hashes are within 6.25% similar,
which is very similar out of the 2^64 combinations of hashes.